code.cpp
It is a mini banking system that stores the data of the user in their system by using file handling applications and allows the user to know what is going on with their bank account. It is like their virtual bank on their laptop. Can be used as a piggy bank manager for managing the data.